working orbit

+++ include/gnclib/Physics.hpp
@@ -7,8 +7,8 @@ using Eigen::Matrix;
 
 template <int svRows> class Physics {
 public:
-  typedef Matrix<float, svRows, 1> StateVector;
-  virtual void eom(float t, const StateVector *x, StateVector *xdot) = 0; // {}
+  typedef Matrix<double, svRows, 1> StateVector;
+  virtual void eom(double t, const StateVector *x, StateVector *xdot) = 0; // {}
 };
 
 #endif // PHYSICS_H
